Moved Data/projects into a new folder inside of projects, and moved Tags into it's own folder in projects
