+++ src/Components/Tag.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import CapitalizeWords from "../functions/CapitalizeWords";
 
 import { Link } from 'react-router-dom';
-import { Language, Tool, System, Type } from "../Data/module";
+import { Language, Tool, System, Type } from "../Data/Tags/module";
 
 export function TagItem({name, to}: {name: Language|Tool|System|Type, to?: string}): JSX.Element {
     const Clickable = to ? Link : "a";
