+++ src/Data/Projects/module.ts
@@ -1,16 +1,10 @@
 import Projects from "./Projects";
 import IProject from "./IProject";
 import StateEnum from "./StateEnum";
-import type { default as Tag, Language, Tool, System, Type } from "./Tags";
 
 export default Projects;
 export {
     Projects,
     IProject,
     StateEnum,
-    Tag,
-        Language,
-        Tool,
-        System,
-        Type
 };
