+++ src/Data/Tags/module.ts
@@ -1,0 +1,10 @@
+import type { default as Tag, Language, Tool, System, Type } from "./Tags";
+
+//export default Projects;
+export {
+    Tag,
+        Language,
+        Tool,
+        System,
+        Type
+};
